Fix warnings

+++ source/Maze.cpp
@@ -2,7 +2,7 @@
 
 #include <random>
 
-Maze::Maze(const int& size)
+Maze::Maze(const size_t& size)
     : m_wallsX{size, std::vector<bool>(size + 1, -1)}
     , m_wallsY{size + 1, std::vector<bool>(size, -1)}
     , m_size{size}
@@ -29,9 +29,9 @@ Maze::Maze(const int& size)
 std::vector<Wall> Maze::walls() const
 {
     std::vector<Wall> result;
-    for (int x = 0; x < size(); ++x)
+    for (size_t x = 0; x < size(); ++x)
     {
-        for (int y = 0; y < size() + 1; ++y)
+        for (size_t y = 0; y < size() + 1; ++y)
         {
             if (isWall(x, y, horizontal))
             {
@@ -39,9 +39,9 @@ std::vector<Wall> Maze::walls() const
             }
         }
     }
-    for (int x = 0; x < m_wallsY.size(); ++x)
+    for (size_t x = 0; x < m_wallsY.size(); ++x)
     {
-        for (int y = 0; y < m_wallsY[x].size(); ++y)
+        for (size_t y = 0; y < m_wallsY[x].size(); ++y)
         {
             if (isWall(x, y, vertical))
             {
@@ -65,7 +65,8 @@ bool Maze::isWall(const int& x, const int& y, Direction direction) const
         return false;
     }
     auto& walls = (direction == horizontal) ? m_wallsX : m_wallsY;
-    if (x < 0 || y < 0 || x >= walls.size() || y >= walls[0].size())
+    if (x < 0 || y < 0 || x >= static_cast<int>(walls.size()) ||
+        y >= static_cast<int>(walls[0].size()))
     {
         return false;
     }
@@ -75,9 +76,9 @@ bool Maze::isWall(const int& x, const int& y, Direction direction) const
 std::vector<Wall> Maze::wallEnds() const
 {
     std::vector<Wall> result;
-    for (int y = 0; y < size() + 1; ++y)
+    for (size_t y = 0; y < size() + 1; ++y)
     {
-        for (int x = 0; x < size() + 1; ++x)
+        for (size_t x = 0; x < size() + 1; ++x)
         {
             if (hasWallAround(x, y))
             {
@@ -107,7 +108,7 @@ Maze::~Maze()
 {
 }
 
-int Maze::size() const
+size_t Maze::size() const
 {
     return m_size;
 }
